Update Astar_proto dir : Completed Subway paths
